Add location tag in modal
